Tap on another ring and gets its status
This patch adds a TapRing() API to ringpop which
queries the given host to get the status of it's ring
dynamically (this is a pull based model rather than a push model
and should generate n/w traffic only on demand).

This will be very useful for client applications which doesn't
want to participate on the gossip protocol and gives them the
ability to get the status of the other ring to make forwarding
decisions.