Change options pattern for Ringpop constructor
This is a breaking change which makes future options changes easier.

Instead of taking constructor options as arguments (or in a struct) we
use a modifier pattern, where each option is a function which modifies
the object being created.

This allows certain options to be ... optional.

This change is to support a future change which will make the "address"
argument optional in the constructor.

Since this is a breaking change, we also take the opportunity to rename
the constructor to ringpop.New, which is more idiomatic.