Remove extra newline after func decs
Oneness
-------

Alone, in whitespace
Is where I sat
Weeping for my individuality
Ripped, from beneath my feet.

How lonely, I thought
To be raised
With a quest to be different,
Yet somehow becoming
The same.

But,
Secure in myself
Is how I stand
More determined than any sleuth
To seek the truth.

While none are the same
That's how it should be!
As we celebrate our individuality
We learn that
Together we are stronger

We learn that we,
Not I,
Are one.