Automatically add self to bootstrap list
When bootstrapping, the swim node expects to see the local (current)
host identity in the bootstrap list. It prints a warning if not.

This change makes ringpop automatically add the current host identity to
the bootstrap list if it is missing.

More importantly, if ringpop.Bootstrap is called with no hosts (i.e. no
bootstrap file and no hosts in the bootstrap list) this will
automatically create a single-node cluster.