add ‘setup’ target to make, update README to match
setup installs the go deps, and checks that thrift is installed,
failing/warning if not.

Changes on master mean that the GOPATH is augmented with the Godeps
workspace. This breaks `go get` by default.

The fix is to invoke `go get` to operate on the user's Go workspace
by resetting GOPATH to the user's default workspace.