Implemented reaping faulty nodes.
Reaping happens via a new state called `tombstone`. This state is ommitted during checksum calculation because the list has the same meaning with and without the tombstones in there. When a node receives a tombstone update for an unknown member it ignores the message (does not add the member to the memberlist, and does not gossip the tombstone message further).