Emit lookup and lookupn stat and add a lookupn-event  (#152)
Previously the lookup stat was emitted from the event-handler inside of ringpop.go when a LookupEvent was emitted. However, the lookup-event is emitted on ringpop itself so it's internal event-handler isn't triggered.

Also a stat and an event on ringpop for LookupN has been added to be on par with ringpop-node.