[Feature] Identity carry over part 3 / Hashring abstraction (#190)
This PR is part 3 for identity carry over.

This PR hooks up the hashring to the newly introduced `membership` package. Instead of talking about servers as strings we respond to `memberhip.MemberChange` slices. This has the benefit that we have access to more fields exposed on a member than just its address.