[Feature] Identity carry over part 4 / Identity carry over (#191)
This PR is part 4 of identity carry over.

This PR actually adds the meat of the story with actual identity carry over. It extends the replica point with more information that will later be used by a new checksumming mechanism and a mechanism that will be backwards compatible with the current checksumming mechanism but added identity support.

Identities are stored in ringpop labels, this mechanism allows for reliable synchronization within the cluster. It uses a key in the private namespace of labels `__identity`. This prevent applications from altering it outside of the supported channels.

If the identity key is missing it will fall back to the address of the Member. This allows for backwards compatibility with other versions and an easy upgrade path when identity is not set on a member.